Back end historico de pedidos, botão cancelar

+++ src/main/java/br/com/rd/ecommerce/model/entity/Pedido.java
@@ -16,7 +16,7 @@ public class Pedido {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cod_pedido")
-    private Integer codPedido;
+    private Long codPedido;
 
     @Column(name = "dt_pedido")
     @Temporal(TemporalType.TIMESTAMP)
